requirements.txt
googlesearch_python==1.2.3
groq==0.13.1                                                                                                                                            
google-generativeai==0.8.3
numpy==1.26.4
pandas==2.2.2
pandas_ta==0.3.14b0
plotly==5.19.0
streamlit==1.41.0
yfinance==0.2.61